Various fixes
Hadn't really run the updated scripts on a clean environment and as such there were a few things that were broken.

+++ fedora/kubernetes-tools.sh
@@ -1,28 +1,23 @@
 #!/bin/bash -e
 
+if [ "$EUID" -eq 0 ]
+  then echo "Please don't run as root"
+  exit
+fi
+
 # Helm
 echo "Installing Helm"
-curl -o helm.tar.gz https://get.helm.sh/helm-v3.0.0-rc.2-linux-amd64.tar.gz
-tar zxvf helm.tar.gz
+curl -o helm.tar.gz https://get.helm.sh/helm-v3.0.0-rc.2-linux-amd64.tar.gz tar zxvf helm.tar.gz
 cp linux-amd64/helm ~/.local/bin
 rm helm.tar.gz
 rm -rf linux-amd64
+helm repo add stable https://kubernetes-charts.storage.googleapis.com
 
 # Kubectl
 echo "Installing Kubectl"
-cat <<EOF > /etc/yum.repos.d/kubernetes.repo
-[kubernetes]
-name=Kubernetes
-baseurl=https://packages.cloud.google.com/yum/repos/kubernetes-el7-x86_64
-enabled=1
-gpgcheck=1
-repo_gpgcheck=1
-gpgkey=https://packages.cloud.google.com/yum/doc/yum-key.gpg https://packages.cloud.google.com/yum/doc/rpm-package-key.gpg
-EOF
-dnf install -y kubectl
+sudo ./kubectl.sh
 
 # Install Minikube
-
 echo "Installing Minikube"
 curl -Lo minikube https://storage.googleapis.com/minikube/releases/v0.32.0/minikube-linux-amd64
 chmod +x minikube
@@ -44,13 +39,16 @@ rm kustomize
 # Install Kind
 go get -u sigs.k8s.io/kind
 
+# Kubectx
 echo "Installing Kubectx"
-git clone https://github.com/ahmetb/kubectx /opt/kubectx
+if [ ! -d /opt/kubectx ]; then
+  sudo git clone https://github.com/ahmetb/kubectx /opt/kubectx
+fi
 
-sudo -u $SUDO_USER zsh << EOF
-  sudo ln -s /opt/kubectx/kubectx ~/.local/bin/kubectx
-  sudo ln -s /opt/kubectx/kubens ~/.local/bin/kubens
-  sudo ln -s /opt/kubectx/completion/kubectx.zsh ~/.oh-my-zsh/completions/_kubctx.zsh
-  sudo ln -s /opt/kubectx/completion/kubens.zsh ~/.oh-my-zsh/completions/_kubens.zsh
-  rm -f ~/.zcompdump;
-EOF
+sudo ln -s /opt/kubectx/kubectx ~/.local/bin/kubectx
+sudo ln -s /opt/kubectx/kubens ~/.local/bin/kubens
+
+mkdir -p ~/.oh-my-zsh/completions
+sudo ln -s /opt/kubectx/completion/kubectx.zsh ~/.oh-my-zsh/completions/_kubctx.zsh
+sudo ln -s /opt/kubectx/completion/kubens.zsh ~/.oh-my-zsh/completions/_kubens.zsh
+rm -f ~/.zcompdump;
